Implemented new Light and Shadowcaster Components
Implemented new PointLight, DirectionalLight and SpotLight Components to replace the exiting LightComponent. Also implemented a new ShadowCasterComponent which stores info specific to light casting shadows such as shadowmap resolution.

+++ VulkanTutorial/src/UI/UIManager.cpp
@@ -56,7 +56,7 @@ namespace Puffin::UI
 		//ImPlot::DestroyContext();
 	}
 
-	void UIManager::DrawUI(float dt)
+	void UIManager::DrawUI(double dt)
 	{
 		ImGui_ImplVulkan_NewFrame();
 		ImGui_ImplGlfw_NewFrame();
